+++ Package.swift
@@ -6,8 +6,9 @@ var dependencies: [Package.Dependency] = [
     .package(url: "https://github.com/jpsim/Yams", from: "6.0.0"),
     .package(url: "https://github.com/tadija/AEXML", from: "4.0.0"),
     .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
-    .package(url: "https://github.com/kateinoigakukun/swift-indexstore", from: "0.4.0"),
-    .package(url: "https://github.com/apple/swift-syntax", from: "602.0.0"),
+    // Use tag once https://github.com/MobileNativeFoundation/swift-index-store/issues/27 is resolved.
+    .package(url: "https://github.com/ileitch/swift-index-store", revision: "ed1f232d33b8e03956af0f4206fbd30171a43138"),
+    .package(url: "https://github.com/apple/swift-syntax", from: "603.0.0"),
     .package(url: "https://github.com/ileitch/swift-filename-matcher", from: "2.0.0"),
 ]
 
@@ -72,7 +73,7 @@ var targets: [PackageDescription.Target] = [
             .product(name: "AEXML", package: "AEXML"),
             .product(name: "SwiftSyntax", package: "swift-syntax"),
             .product(name: "SwiftParser", package: "swift-syntax"),
-            .product(name: "SwiftIndexStore", package: "swift-indexstore"),
+            .product(name: "IndexStore", package: "swift-index-store"),
             .product(name: "FilenameMatcher", package: "swift-filename-matcher"),
         ]
     ),
@@ -81,7 +82,7 @@ var targets: [PackageDescription.Target] = [
         dependencies: [
             .target(name: "SyntaxAnalysis"),
             .target(name: "Shared"),
-            .product(name: "SwiftIndexStore", package: "swift-indexstore"),
+            .product(name: "IndexStore", package: "swift-index-store"),
             .product(name: "AEXML", package: "AEXML"),
         ]
     ),