+++ Sources/SourceGraph/Elements/Declaration.swift
@@ -223,7 +223,7 @@ public final class Declaration {
     public var modifiers: Set<String> = []
     public var accessibility: DeclarationAccessibility = .init(value: .internal, isExplicit: false)
     public let kind: Kind
-    public var name: String?
+    public let name: String
     public let usrs: Set<String>
     public var unusedParameters: Set<Declaration> = []
     public var declarations: Set<Declaration> = []
@@ -235,6 +235,7 @@ public final class Declaration {
     public var related: Set<Reference> = []
     public var isImplicit: Bool = false
     public var isObjcAccessible: Bool = false
+    public var referencedFiles: Set<SourceFile>
 
     private let hashValueCache: Int
 
@@ -290,10 +291,17 @@ public final class Declaration {
         related.filter { $0.declarationKind == kind && $0.name == name }
     }
 
-    public init(kind: Kind, usrs: Set<String>, location: Location) {
+    public init(
+        name: String,
+        kind: Kind,
+        usrs: Set<String>,
+        location: Location
+    ) {
+        self.name = name
         self.kind = kind
         self.usrs = usrs
         self.location = location
+        self.referencedFiles = [location.file]
         hashValueCache = usrs.hashValue
     }
 
@@ -322,7 +330,7 @@ extension Declaration: CustomStringConvertible {
     }
 
     private var descriptionParts: [String] {
-        let formattedName = name != nil ? "'\(name!)'" : "nil"
+        let formattedName = "'\(name)'"
         let formattedAttributes = "[" + attributes.sorted().map(\.description).joined(separator: ", ") + "]"
         let formattedModifiers = "[" + modifiers.sorted().joined(separator: ", ") + "]"
         let formattedCommentCommands = "[" + commentCommands.map(\.description).sorted().joined(separator: ", ") + "]"