+++ Sources/SourceGraph/SourceGraphMutatorRunner.swift
@@ -49,6 +49,7 @@ public final class SourceGraphMutatorRunner {
         PropertyWrapperRetainer.self,
         ResultBuilderRetainer.self,
         CodablePropertyRetainer.self,
+        EquatableHashablePropertyRetainer.self,
         ExternalOverrideRetainer.self,
 
         AncestralReferenceEliminator.self,